changed evaluation, added inverse fitness prop. selection

+++ src/main/kotlin/evo/Individual.kt
@@ -2,6 +2,6 @@ package evo
 
 data class Individual(var data: List<List<Int>>) {
 
-    var fitness: Int = 0
+    var fitness: Double = 0.0
 
 }
